refactor(web-serial-data-access): drop SerialPromptDetectorService
